[DEVELOPMENT ENVIRONMENT] * BUG: Do not abort if you try to open a dead symbolic link.
[INTERPRETER]
* BUG: Fix base object structure size when compiling with gcc 4.3 on a 64 
  bits system.

[INFORMER]
* BUG: The informer now can be run without argument from anywhere.

[GB.FORM]
* NEW: The ListContainer refreshes itself before raising the Click event.


git-svn-id: svn://localhost/gambas/trunk@1481 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/gbx/gbx_object.h
@@ -32,7 +32,7 @@
 typedef
   struct {
     CLASS *class;
-    int ref;
+    intptr_t ref;
     }
   OBJECT;
